Delete existing login before creating a new login

+++ src/lib/actions/Session.js
@@ -34,9 +34,7 @@ function signOut() {
     if (!credentials || !credentials.login) {
         return;
     }
-    API.deleteLogin({
-        partnerUserID: credentials.login
-    });
+    API.deleteLogin({partnerUserID: credentials.login});
 }
 
 export {
